Adding plays

+++ src/controller/PlaysManager.java
@@ -3,5 +3,5 @@ package controller;
 import entities.Song;
 
 public interface PlaysManager {
-    void addPlay(long secondsPlayed, Song song);
+    void addPlay(int secondsPlayed, Song song);
 }
